Move Bisection tests to own file
